+++ CHANGELOG.md
@@ -1,6 +1,13 @@
+## 0.6.1
+
+### Enhancements
+- Added helper methods to simplify setting up simulations
+  - `setup_replica_exchange`
+  - `setup_data_store`
+
 ## 0.6.0
 
-## Bug Fixes
+### Bug Fixes
 - Previous versions of MELD could give erroneous values for forces and energies
   - Due to a data race in the CUDA kernels, different threads could calculate
     different values for the cutoff energy within each collection. This would
@@ -14,7 +21,7 @@
 - Previous versions could sometimes incorrectly negotiate device ids between
   nodes which would lead to failure at runtime.
 
-## Enhancements
+### Enhancements
 - Amber-specific system buiding tools are in their own modules
 - RDCs no longer add particles to the system
 - Energies are saved to different energy groups. The current mapping is:
@@ -53,7 +60,7 @@
     atom names.
   - "follower" changed to "worker"
 
-## Changes
+### Changes
 - No longer use network logging.
   - Previous versions used network logging to aggregate all log messages into a
     single remd.log file.